language/search_agents/requirements.txt
absl-py
apache_beam
attrs
cloudpickle==1.3.0
grpcio>=1.32.0
grpcio-tools
gym
keras
nltk
numpy
pygtrie
tensorflow==2.4.1
tensorflow-addons
tensorflow-probability==0.11.0
tensorflow-serving-api
tensorflow_text
transformers
